Use opcode as first field in packet
Previously, ack's from the component weren't putting 0xff (bad opcode)
into the opcode field of the packet. When packets aren't cleared in
the AP, we could still detect that some packet was received. authTag
checks would eventually tell us that it's an invalid packet, but hey,
this is more accurate handling.